requirements.txt
opencv-python
mediapipe
pyautogui
Pillow
numpy==1.21.0